+++ src/main/java/movieapp/webmovie/service/impl/UserServiceImpl.java
@@ -75,7 +75,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void updateUserRole(Integer userId, String newRole) {
+    public void updateUserRole(Long userId, String newRole) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy người dùng"));
